+++ app/src/main/java/woowacourse/movie/ui/completed/CompletedActivity.kt
@@ -7,9 +7,9 @@ import android.view.MenuItem
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import woowacourse.movie.R
-import woowacourse.movie.domain.Ticket
 import woowacourse.movie.formatScreenDateTime
 import woowacourse.movie.ui.MovieService
+import woowacourse.movie.ui.uiModel.Ticket
 import woowacourse.movie.util.getParcelable
 
 class CompletedActivity : AppCompatActivity() {
