+++ app/src/test/java/woowacourse/movie/domain/TicketTest.kt
@@ -2,6 +2,7 @@ package woowacourse.movie.domain
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
+import woowacourse.movie.ui.uiModel.Ticket
 import java.time.LocalDateTime
 
 class TicketTest {
